# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DineFlow restaurant management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer/tables/{code}:
    get:
      operationId: getTableByCode
      tags: [customer]
      summary: Look up a table by its QR code
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Table info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Table" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /customer/orders:
    post:
      operationId: createCustomerOrder
      tags: [customer]
      summary: Place a new order from a table
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /customer/orders/by-session/{sessionId}:
    get:
      operationId: getOrdersBySession
      tags: [customer]
      summary: List all orders placed in this customer session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /customer/orders/{id}:
    get:
      operationId: getCustomerOrder
      tags: [customer]
      summary: Get an order with bill breakdown
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /customer/orders/{id}/call-waiter:
    post:
      operationId: callWaiter
      tags: [customer]
      summary: Send a call-waiter alert from a table
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /customer/orders/{id}/pay:
    post:
      operationId: payOrder
      tags: [customer]
      summary: Pay for an order (mock payment)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayOrderInput" }
      responses:
        "200":
          description: Paid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }

  # ─────────────── AUTH ───────────────
...
components:
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }

    # ─────────────── DOMAIN ───────────────
...
    Table:
      type: object
      required: [id, code, label, capacity, status, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        label: { type: string }
        capacity: { type: integer }
        status:
          type: string
          enum: [available, seated, dining, billing, cleaning]
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
...
    OrderItem:
      type: object
      required: [id, menuItemId, name, quantity, unitPrice, lineTotal]
      properties:
        id: { type: integer }
        menuItemId: { type: integer }
        name: { type: string }
        imageEmoji:
          type: ["string", "null"]
        quantity: { type: integer }
        unitPrice: { type: number }
        lineTotal: { type: number }
        specialInstructions:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, code, tableId, tableLabel, sessionId, status, items, subtotal, tax, tip, total, createdAt, updatedAt]
      properties:
        id: { type: integer }
        code: { type: string }
        tableId: { type: integer }
        tableLabel: { type: string }
        sessionId: { type: string }
        status:
          type: string
          enum: [placed, preparing, ready, served, paid, cancelled]
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        subtotal: { type: number }
        tax: { type: number }
        tip: { type: number }
        total: { type: number }
        guestCount:
          type: ["integer", "null"]
        guestName:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateOrderBody:
      type: object
      required: [tableCode, sessionId, items]
      properties:
        tableCode: { type: string }
        sessionId: { type: string }
        guestName:
          type: ["string", "null"]
        guestCount:
          type: ["integer", "null"]
        items:
          type: array
          items:
            type: object
            required: [menuItemId, quantity]
            properties:
              menuItemId: { type: integer }
              quantity: { type: integer }
              specialInstructions:
                type: ["string", "null"]
...
    PayOrderInput:
      type: object
      required: [method, tip]
      properties:
        method:
          type: string
          enum: [upi, card, cash]
        tip: { type: number }

    # ─────────────── AUTH ───────────────
...